refactored file structure.
