+++ lib/Controllers/home_screen_controller.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:podboi/Services/database/settings_box_controller.dart';
+import 'package:podboi/Database/settings_box_controller.dart';
 import 'package:podcast_search/podcast_search.dart';
 
 var _search = Search();
