+++ lib/Controllers/settings_controller.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:podboi/Services/database/settings_box_controller.dart';
+import 'package:podboi/Database/settings_box_controller.dart';
 
 final settingsController =
     StateNotifierProvider<SettingsStateNotifier, SettingsState>((ref) {
